go for typscript
