+++ .gitignore
@@ -1,3 +1,5 @@
 build
 /js/vendor/
 nbproject
+
+*/**.js
