+++ js/OC.ts
@@ -1,0 +1,13 @@
+/* global OC */
+
+/**
+ * ownCloud - Sketch
+ *
+ * This file is licensed under the Affero General Public License version 3 or
+ * later. See the COPYING file.
+ *
+ * @author Christoph Wurst <christoph@winzerhof-wurst.at>
+ * @copyright Christoph Wurst 2015
+ */
+
+export = OC;
