+++ js/start.ts
@@ -1,5 +1,3 @@
-/* global OC */
-
 /**
  * ownCloud - Sketch
  *
@@ -7,11 +5,13 @@
  * later. See the COPYING file.
  *
  * @author Christoph Wurst <christoph@winzerhof-wurst.at>
- * @copyright Christoph Wurst 2015
+ * @copyright Christoph Wurst 2016
  */
 
-define(function() {
-	'use strict';
+import app = require('App');
 
-	return OC;
-})
+app.on('start', function() {
+	app.Controller.SketchController.loadAll();
+});
+
+app.start();
