+++ js/tsconfig.json
@@ -1,0 +1,10 @@
+{
+	"compilerOptions": {
+		"target": "ES5",
+		"module": "amd",
+		"sourceMap": false
+	},
+	"exclude": [
+		"vendor"
+	]
+}
