+++ js/view/SketchListItem.ts
