+++ typings.json
@@ -1,0 +1,5 @@
+{
+  "name": "sketch",
+  "dependencies": {},
+  "devDependencies": {}
+}
